requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
requests==2.31.0
openai==1.3.7
SpeechRecognition==3.10.0
pydub==0.25.1

